feat: Implement MVP for VQuest game
This commit introduces the minimum viable product for the VQuest game.
It includes the following changes:

- **Core Game Loop:** Implemented the core game loop on the server, including advancing the game state, selecting a winning action, and broadcasting updates to clients.
- **AI Dungeon Master:** Integrated a basic AI Dungeon Master using the Eden AI API to generate story content based on player actions.
- **Player View:** Improved the player view to include a WebSocket connection for live updates, display of game state (context, prompt, actions), and voting functionality.
- **DM View:** Improved the DM view to include a "Next Step" button to advance the game state.
- **Bug Fixes & Refactoring:**
  - Fixed duplicated code in `static/player.html`.
  - Refactored the server to separate the Express app from the server logic, enabling better testing.
  - Fixed path issues for static assets.
- **Testing:** Added a basic test suite for the server-side logic.
